Allow image editing
Allow editing of images. photofs transparently forwards the write
calls to the real files.

Implemented photo update routine so it's no longer necessary to
delete and then add.

Added a hack to allow exiv2 library to work. It wants to create a
temp file, write into it and then rename into the original. The hack
fools it by creating a file in /tmp and operating on that.

Bump version number.